#192 Move cache classes to separate packages
