+++ action/library/src/main/java/io/knotx/fragments/action/library/cache/CacheStore.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.fragments.action.library;
+package io.knotx.fragments.action.library.cache;
 
 import io.knotx.fragments.api.FragmentResult;
